# Conflicts:
#	src/main/java/com/fitbum/servicios/mapper/UsuarioMapper.java

+++ src/main/java/com/fitbum/repositorios/estadisticas/EstadisticasRepositorio.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 
-public interface EstadisticasRepositorio extends JpaRepository<Estadisticas,Long> {
+public interface EstadisticasRepositorio extends JpaRepository<Estadisticas,Integer> {
 
     Page<Estadisticas> findAll(Pageable pageable);
 }
