+++ src/main/java/com/fitbum/repositorios/tutoriales/ContentTutorialRepositorio.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ContentTutorialRepositorio  extends JpaRepository<ContentTutorial,Long> {
+public interface ContentTutorialRepositorio  extends JpaRepository<ContentTutorial,Integer> {
 }
